[MI-3091] Separated the functionality to create sunscription of new incident inn different react component
